Refactor: centralize scan task logic in ScanTaskBase
Extract shared lock, state, and execution logic from BulkScanTaskBase and DailyScanTaskBase into a new abstract base class, ScanTaskBase. Test code and constructors are updated accordingly.
